don't initialize new oval Validator on every validation check
This caused a memory leak in conjunction with groovy expressions in oval validators (when="grovy:...").
Every time a new validator is created an such an expression must be evaluated, groovy was registered as new expression language.
The expression language classloader and the loaded class of the expression never got garbage collected which caused permgen space errors.